% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gss14_f}
\alias{gss14_f}
\title{GSS 2014 select variables as factors.}
\format{\describe{A data frame with 2538 rows and 4 variables:
  \item{bible}{Factor, feelings about the bible.}
  \item{relig}{Factor, respondents religious preference.}
  \item{denom}{Factor, specific denomination.}
  \item{other}{Factor, other Protestant denominations.}
}}
\source{
\href{http://www.gss.norc.org/get-the-data/spss}{GSS 2014}
}
\usage{
gss14_f
}
\description{
Four varaibles from US Genearl Social Survey from 2014, three describe
respondent's religious identification(\code{relig}, \code{denom}, \code{other}),
and fourth \code{bible} is for biblical litteracy, or answers on the question
if the bible is "the word of god", "inspired word", "book of fables", or "other".
Data is used in vignettes, to demonstrate the use of
\code{\link{recode_religion}}, and to provide some comparison between
the use of packages \code{foreign} and \code{haven} in importing SPSS data
and their different recoding of categorical variables from GSS. \code{gss14_f}
was imported through \code{foreign}, with use of following options:
\code{to.data.frame = TRUE, trim.factor.names = TRUE, trim_values = TRUE,
use.missings = FALSE}, accordingly varaibles are factors without missing
values.
}
\seealso{
\code{\link[foreign]{read.spss}} \code{\link[haven]{read_sav}}
}
\keyword{datasets}
