% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gss14_n}
\alias{gss14_n}
\title{GSS 2014 select variables as numerical.}
\format{\describe{A data frame with 2538 rows and 4 variables:
  \item{bible}{Labelled num, feelings about the bible.}
  \item{relig}{Labelled num, respondents religious preference.}
  \item{denom}{Labelled num, specific denomination.}
  \item{other}{Labelled num, other Protestant denominations.}
}}
\source{
\href{http://www.gss.norc.org/get-the-data/spss}{GSS 2014}
}
\usage{
gss14_n
}
\description{
Four varaibles from US Genearl Social Survey from 2014, three describe
respondent's religious identification(\code{relig}, \code{denom}, \code{other}),
and fourth \code{bible} is for biblical litteracy, or answers on the question
if the bible is "the word of god", "inspired word", "book of fables", or "other".
Data is used in vignettes, to demonstrate the use of
\code{\link{recode_religion}}, and to provide some comparison between
the use of packages \code{foreign} and \code{haven} in importing SPSS data
and their different recoding of categorical variables from GSS. \code{gss14_n}
was imported through \code{haven}, and thus variables are specific numeric
\code{labelled} variables, with attributes "label" for description,
"format.spss" for SPSS format, and "labels" which are names for levels.
}
\seealso{
\code{\link[haven]{read_sav}} \code{\link[foreign]{read.spss}}
}
\keyword{datasets}
