% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rdo_cdbk}
\alias{rdo_cdbk}
\title{Codebook for relig, denom, and other variables.}
\format{A list containing three data frames:
\describe{
  \item{relig}{data frame with 16 observations and 2 variables:
  \code{punch}, numeric; \code{label} factor.}
  \item{denom}{data frame with 30 observations and 2 variables:
  \code{punch}, numeric; \code{label} factor.}
  \item{other}{data frame with 201 observation and 2 variables:
  \code{punch}, numeric; \code{label} factor.}
}}
\source{
\href{http://www.gss.norc.org/Get-Documentation}{General Social Survey}
}
\usage{
rdo_cdbk
}
\description{
A list containing data frames that have individual punches and
labels for \code{relig}, \code{denom}, and \code{other} variables, which are also
applicable for whole family of variables such as religious
identification at the age of 16, spouses, parents, etc. in General
Social Survey from 1972 until 2014. Some of these non-primariy
variables are often coded only with punches, and not with labels.
It's used in \code{\link{rdo_punches}}.
}
\keyword{datasets}
