% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{rdo_punches}
\alias{rdo_punches}
\title{Pulls punches and labels for religious identification variables.}
\usage{
rdo_punches(vrbl)
}
\arguments{
\item{vrbl}{Character string either 'relig', 'denom', or 'other'.}
}
\value{
Character vector with religion from either \code{relig},
\code{denom}, or \code{other} family of variables, where
index coresponds to punch code and value to label. Different
missing values are described in coresponding values.
}
\description{
Function that returns vector with punches and
labels for religious identification family of variables.
Since, many punches are skipped, especially in \code{other},
it has a lot of \code{NA}, while punches for different
missing values are described, e.g. \code{'NA'}. Does not return
punch for \code{'IAP'}, since it is 0. It's used in
\code{\link{fct_rec_relig}}
}
\examples{
relig_punches <- rdo_punches("relig")
}
