% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_religion.R
\name{recode_religion}
\alias{recode_religion}
\title{Recode Religion from GSS.}
\usage{
recode_religion(relig, denom, other, n_groups = 12,
  add_missing_levels = FALSE, frequencies = TRUE, print_key = FALSE,
  return_num = FALSE)
}
\arguments{
\item{relig, denom, other}{Numerical, character, or factor, all of same
length and with coresponding punches or labels in codebook.}

\item{n_groups}{Number 12, i.e. number of new religious identifications.}

\item{add_missing_levels}{Logical, to include as empty levels religious
identifications that may not be present in specific sample, but are
part of recoding schema.}

\item{frequencies}{Logical, to print frequency and percent table of
recoded religius identification (default is \code{TRUE}).}

\item{print_key}{Logical, to print the all unique tetrads of recoded
variables, i.e. a recoding key.}

\item{return_num}{Logical, to return numerical factor and print codebook.}
}
\value{
Vector with recoded religion from \code{relig}, \code{denom},
and \code{other}. Function does not return \code{NA}, but as
factor levels "Not answered" and "Don't know", or combined
"Not answered/Don't know" when missing values are not
declared as punches or labels in initial variables but passed on as
\code{NA} (function gives message and where \code{NAs} are lcoated).
Default is to have factor with 12 descriptive levels, but function can
also return numerical vector. Default behavior returns only present
values, but can be made to add additional empty levels if \code{TRUE}
is passed to \code{add_missing_levels}. Function also \strong{prints}
frequency table of newly recoded religious identification, which can be
suppressed with \code{frequencies}. If required, it can also return
numerical vector and print the coding for it (not recommended).
}
\description{
Function recodes religious identification from Genearl Social Survey,
based on three variables: \code{relig}, \code{denom}, and \code{other}.
It can successfully recode either respondent's, or any other religious
identification which is determened by coresponding three variables.
}
\details{
\code{recode_religion} uses schema developed by Darren E. Sherkat and
Derek Lehman in \href{https://iranianredneck.wordpress.com/2016/11/29/why-reltrad-sucks-contesting-the-measure-of-american-religion/}{"After
The Resurrection: The Field of the Sociology of Religion in the United
States"}, and is effectievly translation of that SPSS syntax (the bare
bone function for recoding is \link{fct_rec_relig}), with
additional functionality.

Namley, it can handle both punches and labels
at the same time (but in different variables), which is important since
punches are not consequtive as indexes. In addition, function checks that
variables are adequate (i.e. that all values are in codebook) and of same
length, and also handles missing values: (1) if supplied through values,
provides detail recoding; (2) if \code{NA}, lumps them together in final
variable but uses them correctly in the recoding. Through passed
arguments, one can:
\enumerate{
  \item Add identifications from schema that are not present in
  sample as empty levels.
  \item Suppress printing of the frequencies of newly recoded variable.
  \item Print unique key of the values that were recoded.
  \item Return values as numerical factor, in which case the codebook
  for new variable will be printed.
}
If \code{frequencies} is passed as \code{FALSE}, and numerical vector
is not requested as return value, all other information, such as
treatment of missing values, are provided as messages that can be
suppressed.

Future behavior will provide recoding to 7 levels. More details can be
found on \href{https://github.mdjeric}{github.mdjeric}.
}
\examples{
library(resurrectionr)

# When all variables are factor
gss14_f$religion <- recode_religion(gss14_f$relig, gss14_f$denom,
                                    gss14_f$other, frequencies = FALSE)

# When all variables are numeric
gss14_n$religion <- recode_religion(gss14_n$relig, gss14_n$denom,
                                    gss14_n$other,
                                    add_missing_levels = TRUE)

# But also, combining them works
religion <- recode_religion(gss14_f$relig, gss14_n$denom,
                            as.character(gss14_f$other))
}
